refactor: agents.langchain directory
