+++ examples/agriculture-demo.py
@@ -18,7 +18,10 @@ import rclpy
 from langchain_core.messages import HumanMessage
 from langchain_core.runnables import Runnable
 from rai import get_llm_model
-from rai.agents.conversational_agent import State, create_conversational_agent
+from rai.agents.langchain.core.conversational_agent import (
+    State,
+    create_conversational_agent,
+)
 from rai.communication.ros2.connectors import ROS2Connector
 from rai.tools.ros2 import ROS2ServicesToolkit, ROS2TopicsToolkit
 from rai.tools.time import WaitForSecondsTool
