+++ src/rai_core/rai/agents/__init__.py
@@ -13,16 +13,13 @@
 # limitations under the License.
 
 from rai.agents.base import BaseAgent
-from rai.agents.conversational_agent import create_conversational_agent
-from rai.agents.langchain.react_agent import ReActAgent
+from rai.agents.langchain import BaseStateBasedAgent, ReActAgent
 from rai.agents.runner import AgentRunner, wait_for_shutdown
-from rai.agents.tool_runner import ToolRunner
 
 __all__ = [
     "AgentRunner",
     "BaseAgent",
+    "BaseStateBasedAgent",
     "ReActAgent",
-    "ToolRunner",
-    "create_conversational_agent",
     "wait_for_shutdown",
 ]
