+++ src/rai_core/rai/agents/langchain/core/runnables.py
@@ -31,7 +31,7 @@ from langchain_core.tools import BaseTool
 from langgraph.graph import START, StateGraph
 from langgraph.prebuilt.tool_node import tools_condition
 
-from rai.agents.tool_runner import ToolRunner
+from rai.agents.langchain.core.tool_runner import ToolRunner
 from rai.initialization import get_llm_model
 from rai.messages import HumanMultimodalMessage, SystemMultimodalMessage
