+++ tests/core/test_tool_runner.py
@@ -16,7 +16,7 @@ import logging
 
 from langchain_core.messages import AIMessage, ToolCall, ToolMessage
 from langchain_core.tools import tool
-from rai.agents.tool_runner import ToolRunner
+from rai.agents.langchain.core import ToolRunner
 from rai.messages import HumanMultimodalMessage, ToolMultimodalMessage, preprocess_image
 from rai.tools.ros2.cli import ros2_topic
